+++ spec/units/gandi_v5_spec.rb
@@ -47,6 +47,10 @@ describe GandiV5 do
   end
 
   describe 'Uses RestClient' do
+    let(:response) do
+      double RestClient::Response, body: 'Hello world!', headers: { content_type: 'text/plain' }
+    end
+
     %i[get delete].each do |method|
       describe ":#{method}" do
         it 'As JSON' do
@@ -58,30 +62,30 @@ describe GandiV5 do
         end
 
         it 'As text' do
-          response_data = 'Hello world!'
-          response = double RestClient::Response, body: response_data, headers: { content_type: 'text/plain' }
           expect(described_class).to receive(:prepare_headers)
           expect(RestClient).to receive(method).with('url', hash_including(accept: 'text/plain')).and_return(response)
-          expect(described_class.send(method, 'url', accept: 'text/plain')).to match_array [response, response_data]
+          expect(described_class.send(method, 'url', accept: 'text/plain')).to match_array [response, 'Hello world!']
         end
 
         it 'Passes request headers' do
           expect(described_class).to receive(:prepare_headers)
           expect(described_class).to receive(:parse_response)
-          expect(RestClient).to receive(method).with(anything, header: 'value')
-          expect(described_class.send(method, 'url', header: 'value')).to match_array [nil, nil]
+          expect(RestClient).to receive(method).with(anything, header: 'value').and_return(response)
+          expect(described_class.send(method, 'url', header: 'value')).to match_array [response, nil]
         end
 
         it 'Adds authentication header' do
           expect(RestClient).to receive(method).with(anything, hash_including(Authorization: 'Apikey abdce12345'))
+                                               .and_return(response)
           expect(described_class).to receive(:parse_response)
-          expect(described_class.send(method, 'https://api.gandi.net/v5/')).to match_array [nil, nil]
+          expect(described_class.send(method, 'https://api.gandi.net/v5/')).to match_array [response, nil]
         end
 
         it 'Default accept header' do
           expect(RestClient).to receive(method).with(any_args, hash_including(accept: 'application/json'))
+                                               .and_return(response)
           expect(described_class).to receive(:parse_response)
-          expect(described_class.send(method, 'https://api.gandi.net/v5/')).to match_array [nil, nil]
+          expect(described_class.send(method, 'https://api.gandi.net/v5/')).to match_array [response, nil]
         end
 
         it 'Converts a 406 (bad request) exception' do
@@ -90,6 +94,13 @@ describe GandiV5 do
           described_class.send(method, 'https://api.gandi.net/v5/')
         end
       end
+    end
+
+    it ':delete handles no content-type' do
+      response = double RestClient::Response, headers: {}
+      expect(described_class).to receive(:prepare_headers)
+      expect(RestClient).to receive(:delete).with('url', hash_including(accept: 'text/plain')).and_return(response)
+      expect(described_class.delete('url', accept: 'text/plain')).to match_array [response, nil]
     end
 
     %i[patch post put].each do |method|
@@ -107,26 +118,25 @@ describe GandiV5 do
         end
 
         it 'As text' do
-          response_data = 'hello world'
-          response = double RestClient::Response, body: response_data, headers: { content_type: 'text/plain' }
           expect(described_class).to receive(:prepare_headers)
           expect(RestClient).to receive(method).with('url', payload, hash_including(accept: 'text/plain'))
                                                .and_return(response)
-          expect(described_class.send(method, 'url', payload, accept: 'text/plain')).to match_array [response, response_data]
+          array = [response, 'Hello world!']
+          expect(described_class.send(method, 'url', payload, accept: 'text/plain')).to match_array array
         end
 
         it 'Passes payload' do
           expect(described_class).to receive(:prepare_headers)
           expect(described_class).to receive(:parse_response)
-          expect(RestClient).to receive(method).with(anything, payload, any_args)
-          expect(described_class.send(method, 'url', payload)).to match_array [nil, nil]
+          expect(RestClient).to receive(method).with(anything, payload, any_args).and_return(response)
+          expect(described_class.send(method, 'url', payload)).to match_array [response, nil]
         end
 
         it 'Passes request headers' do
           expect(described_class).to receive(:prepare_headers)
           expect(described_class).to receive(:parse_response)
-          expect(RestClient).to receive(method).with(any_args, hash_including(header: 'value'))
-          expect(described_class.send(method, 'url', payload, header: 'value')).to match_array [nil, nil]
+          expect(RestClient).to receive(method).with(any_args, hash_including(header: 'value')).and_return(response)
+          expect(described_class.send(method, 'url', payload, header: 'value')).to match_array [response, nil]
         end
 
         it 'Adds content type header' do
@@ -137,14 +147,16 @@ describe GandiV5 do
 
         it 'Adds authentication header' do
           expect(RestClient).to receive(method).with(any_args, hash_including(Authorization: 'Apikey abdce12345'))
+                                               .and_return(response)
           expect(described_class).to receive(:parse_response)
-          expect(described_class.send(method, 'https://api.gandi.net/v5/', payload)).to match_array [nil, nil]
+          expect(described_class.send(method, 'https://api.gandi.net/v5/', payload)).to match_array [response, nil]
         end
 
         it 'Default accept header' do
           expect(RestClient).to receive(method).with(any_args, hash_including(accept: 'application/json'))
+                                               .and_return(response)
           expect(described_class).to receive(:parse_response)
-          expect(described_class.send(method, 'https://api.gandi.net/v5/', payload)).to match_array [nil, nil]
+          expect(described_class.send(method, 'https://api.gandi.net/v5/', payload)).to match_array [response, nil]
         end
 
         it 'Converts a 406 (bad request) exception' do